requirements.txt
httpx==0.27.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.4
pillow==11.0.0
python-dotenv==1.0.1
openai==2.8.1